Fix handling of back button in fragment/activity
The issue was that a fragment was being restored by the framework on
orientation changes. Such a fragment would show nothing since its onStart()
returns if the mediabrowser is not connected. This was compounded by the
fact that onConnected() in the activity was adding a new fragment on each call.

This fix does the following:
1. Moves the code that should be run when a MediaBrowser is connected into
a method named onConnected()
2. Changes the Activity to start with a fragment if savedInstanceState is null.
3. Changes the connectionCallback of the activity to call the fragment.onConnected()
4. Finally it handles the case where a new fragment is added when the browser is already
connected => onStart() called onConnected()

Bug:18791818
Change-Id: I6cb2d5403d95349443e296598cf1bcf09a15e076